Rydde rediger/legg til knapp
